聊天记录分页按钮忘了激活了(项目展示视频忘了展示)

+++ src/main/java/com/melo/wechat/dao/inter/MessageDao.java
@@ -26,7 +26,7 @@ public interface MessageDao extends BaseDao {
      * @date: 18:33 2021/5/25
      * @return: java.lang.Integer
      */
-    Integer getMessageCount(Integer chatId);
+     Long getMessageCount(Integer chatId);
 
     /**
      * @Description: 根据前台页码展示特定页
